implement tweet management; add routes and controller for creating, retrieving, updating, and deleting tweets; integrate Cloudinary for image uploads; update tweet model to handle single image

+++ src/utils/UploadToCloudinary.util.js
@@ -9,6 +9,9 @@ cloudinary.config({
 
 const uploadToCloudinary = async (fileLocalPath) => {
   try {
+    if (!fileLocalPath) {
+      return null;
+    }
     const data = await cloudinary.uploader.upload(fileLocalPath);
     if (data) {
       fs.unlinkSync(fileLocalPath);
